Start an Application class
This is theoretically more testable than having all of this logic inline in
main.py, and it also gives us someplace to move the _mailer.
